Adds granular execution time tracking to ksh
This adds detailed tracking of execution time to all commands in an
interactive ksh instance (and only an interactive version). This is
useful if you want to expose such information to other clients, or even
your shelp prompt ($PS1) configuration.

It basically wraps all commands executed with a call to clock_gettime(3)
for a monotonic clock, which it then uses to track the time spent
executing each command, and then exposes that information in two ways:

1. A new prompt escape sequence '\x' which is expanded to be the
   execution time of the preivously completed command, in a human
   readable format, showing two levels of signifance in measuring the
   time (minutes/seconds/microseconds/milliseconds/nanoseconds). This
   seems like a good default for most cases.
2. A new special variable $NANOSECONDS is exposed, just like $SECONDS,
   but it tracks just the nanoseconds spent in the previously executed
   command. This is useful if somebody wants to further
   control/customize how to display such information.

Testing is ongoing to make sure batch mode and is unaffected (timers
aren't even setup there) and the interactive case Just-Works, with
minimal performance impact (the clock_getttime(3) call for a monotonic
timer just calls nanoupdate(3)).

Further testing and tweaking ongoing.
Feedback welcome.